fix review round 1: move ChatGPT OAuth constants to gents-protocol (#1339)
Per review: CLIENT_ID/REFRESH_TOKEN_URL/REFRESH_TOKEN_URL_OVERRIDE_ENV_VAR
now live in gents_protocol::chatgpt_oauth, not gents-chatgpt-login. Both
gents-chatgpt-login (re-exports them for its own public API and internal
use) and gents::chatgpt_oauth_refresh import from gents-protocol directly.
Removed the gents -> gents-chatgpt-login dependency edge from
crates/gents/Cargo.toml (gents already depended on gents-protocol). The
drift test (asserting REFRESH_TOKEN_URL == "{DEFAULT_ISSUER}/oauth/token")
moved with the constants into gents-protocol::chatgpt_oauth.
