+++ crates/gents-chatgpt-login/Cargo.toml
@@ -8,6 +8,7 @@ description = "Minimal ChatGPT OAuth login flows for Gents"
 
 [dependencies]
 base64.workspace = true
+gents-protocol.workspace = true
 rand.workspace = true
 reqwest.workspace = true
 serde.workspace = true
